# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "API Documentation",
    "version" : "1.0"
  },
  "paths" : {
...
    "/add" : {
      "post" : {
        "summary" : "addCommentLike",
        "description" : "addCommentLike",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        }, {
          "name" : "commentId",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "string",
                      "description" : "响应数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/sysRolePermission/add" : {
      "post" : {
        "summary" : "新增数据",
        "description" : "新增数据",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "id" : {
                    "type" : "integer",
                    "description" : "",
                    "format" : "int64"
                  },
                  "roleId" : {
                    "type" : "integer",
                    "description" : "",
                    "format" : "int64"
                  },
                  "permissionId" : {
                    "type" : "integer",
                    "description" : "",
                    "format" : "int64"
                  }
                },
                "description" : "实体对象"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "新增结果",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "object",
                      "properties" : { },
                      "description" : "响应数据"
                    }
                  },
                  "description" : "新增结果"
                }
              }
            }
          }
        }
      }
    },
    "/sysRolePermission/select" : {
      "get" : {
        "summary" : "根据ID查询单条数据",
        "description" : "根据ID查询单条数据",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "主键ID",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "查询结果",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "string",
                      "description" : "响应数据"
                    }
                  },
                  "description" : "查询结果"
                }
              }
            }
          }
        }
      }
    },
    "/sysRolePermission/update" : {
      "put" : {
        "summary" : "修改数据",
        "description" : "修改数据",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "id" : {
                    "type" : "integer",
                    "description" : "",
                    "format" : "int64"
                  },
                  "roleId" : {
                    "type" : "integer",
                    "description" : "",
                    "format" : "int64"
                  },
                  "permissionId" : {
                    "type" : "integer",
                    "description" : "",
                    "format" : "int64"
                  }
                },
                "description" : "实体对象"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "修改结果",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "object",
                      "properties" : { },
                      "description" : "响应数据"
                    }
                  },
                  "description" : "修改结果"
                }
              }
            }
          }
        }
      }
    },
    "/sysRolePermission/delete" : {
      "delete" : {
        "summary" : "根据ID删除数据",
        "description" : "根据ID删除数据",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "主键ID",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "删除结果",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "object",
                      "properties" : { },
                      "description" : "响应数据"
                    }
                  },
                  "description" : "删除结果"
                }
              }
            }
          }
        }
      }
    },
...
    "/sysRole/add" : {
      "post" : {
        "summary" : "add",
        "description" : "add",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "required" : [ "roleName" ],
                "type" : "object",
                "properties" : {
                  "id" : {
                    "type" : "integer",
                    "description" : "",
                    "format" : "int64"
                  },
                  "roleName" : {
                    "type" : "string",
                    "description" : "角色名"
                  },
                  "roleCode" : {
                    "type" : "string",
                    "description" : "角色标识 ADMIN/USER"
                  },
                  "remark" : {
                    "type" : "string",
                    "description" : "备注"
                  }
                },
                "description" : ""
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "string",
                      "description" : "响应数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sysRole/select" : {
      "get" : {
        "summary" : "getById",
        "description" : "getById",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "string",
                      "description" : "响应数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sysRole/update" : {
      "put" : {
        "summary" : "update",
        "description" : "update",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "required" : [ "roleName" ],
                "type" : "object",
                "properties" : {
                  "id" : {
                    "type" : "integer",
                    "description" : "",
                    "format" : "int64"
                  },
                  "roleName" : {
                    "type" : "string",
                    "description" : "角色名"
                  },
                  "roleCode" : {
                    "type" : "string",
                    "description" : "角色标识 ADMIN/USER"
                  },
                  "remark" : {
                    "type" : "string",
                    "description" : "备注"
                  }
                },
                "description" : ""
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "string",
                      "description" : "响应数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sysRole/delete" : {
      "delete" : {
        "summary" : "delete",
        "description" : "delete",
        "parameters" : [ {
          "name" : "ids",
          "in" : "query",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "array",
            "nullable" : false,
            "items" : {
              "type" : "integer"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "string",
                      "description" : "响应数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sysRole/list" : {
      "get" : {
        "summary" : "list",
        "description" : "list",
        "parameters" : [ {
          "name" : "pageNum",
          "in" : "query",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false,
            "default" : "1"
          }
        }, {
          "name" : "pageSize",
          "in" : "query",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false,
            "default" : "10"
          }
        }, {
          "name" : "roleName",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        }, {
          "name" : "roleCode",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "object",
                      "properties" : {
                        "records" : {
                          "type" : "array",
                          "description" : "",
                          "items" : {
                            "required" : [ "roleName" ],
                            "type" : "object",
                            "properties" : {
                              "id" : {
                                "type" : "integer",
                                "description" : "",
                                "format" : "int64"
                              },
                              "roleName" : {
                                "type" : "string",
                                "description" : "角色名"
                              },
                              "roleCode" : {
                                "type" : "string",
                                "description" : "角色标识 ADMIN/USER"
                              },
                              "remark" : {
                                "type" : "string",
                                "description" : "备注"
                              }
                            },
                            "description" : "(SysRole)表实体类"
                          }
                        },
                        "total" : {
                          "type" : "integer",
                          "description" : "",
                          "format" : "int64"
                        },
                        "size" : {
                          "type" : "integer",
                          "description" : "",
                          "format" : "int64"
                        },
                        "current" : {
                          "type" : "integer",
                          "description" : "",
                          "format" : "int64"
                        },
                        "orders" : {
                          "type" : "array",
                          "description" : "",
                          "items" : {
                            "type" : "object",
                            "properties" : {
                              "column" : {
                                "type" : "string",
                                "description" : ""
                              },
                              "asc" : {
                                "type" : "boolean",
                                "description" : ""
                              }
                            },
                            "description" : "com.baomidou.mybatisplus.core.metadata.OrderItem"
                          }
                        },
                        "optimizeCountSql" : {
                          "type" : "boolean",
                          "description" : ""
                        },
                        "searchCount" : {
                          "type" : "boolean",
                          "description" : ""
                        },
                        "optimizeJoinOfCountSql" : {
                          "type" : "boolean",
                          "description" : ""
                        },
                        "maxLimit" : {
                          "type" : "integer",
                          "description" : "",
                          "format" : "int64"
                        },
                        "countId" : {
                          "type" : "string",
                          "description" : ""
                        }
                      },
                      "description" : "响应数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}